refactor(issue-lifecycle): inline implementations, delete old modules
Inline all implementations from the 7 old issue modules (issues.ts,
label-lifecycle.ts, labels.ts, issue-dispatch.ts, prd-discovery.ts,
prd-hitl.ts, reset-labels.ts) into issue-lifecycle.ts and
issue-naming.ts. Delete the old modules and their test files.
Update all remaining imports and ARCHITECTURE.md.
